This is a sample Exodus II file for the pCAMAL interface. I generated the mesh data with Cubit and added the object properties with another program I wrote.
The model is a thick-walled cylinder split down the middle. The sweep
direction for both halve is along the cylindrical axis. All element
blocks, node set, and sides sets with "SweepID" 1 are part of the first
sweep volume, and those with "SweepID" 2 are the second sweep volume.
Side sets with "SurfaceType" 1 are source, 2 are linking, and 3 are
target quads.

See ./cubit_test/sweep-basic/bisected_cyl.sat for the original model.